requirements.txt
Flask==2.3.3
openpyxl==3.1.2
gunicorn==21.2.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0

